requirements.txt
fastapi
uvicorn[standard]
geopy
pyswisseph
openai